refactor(postgres): split postgres files into their own service group
